Make CPU manager robust to sockets with zNUMAs
The CPU manager uses the number of sockets on the host to compute stuff
such as number of cores per sockets, etc... , which it needs in its
algorithms. Prior to this commit, it considered the number of sockets to
be the number of physical sockets on the host.

In our setup, where a physical socket might hold one zNUMA and nothing
else, that can cause counting errors, since the CPU manager assumes that
all sockets hold the same number of CPUs. So this commit makes the CPU
manager exclude sockets that hold zNUMAs from its computation. Note that
this commit still assumes either a 1-1 mapping between sockets and NUMAs
(l or z), or at least all sockets to be configured the same way in terms
of how many l and zNUMAs they hold. In setups where sockets hold more
than one NUMA, and different sockets have a different breakdown between
l and z NUMAs, this commit is not enough.